requirements.txt
Django~=5.2.10
jwt~=1.4.0
PyJWT~=2.10.1
djangorestframework~=3.16.1
cryptography~=46.0.3
django-stubs~=5.2.8
requests